+++ src/index.ts
@@ -469,6 +469,36 @@ export class Scru128Generator {
   getLastStatus() {
     return this.lastStatus;
   }
+
+  /**
+   * Returns an infinite iterator object that produces a new ID for each call of
+   * `next()`.
+   *
+   * @example
+   * ```javascript
+   * import { Scru128Generator } from "scru128";
+   *
+   * const [a, b, c] = new Scru128Generator();
+   * console.log(String(a)); // e.g. "038MQR9E14CJC12DH9AMW7I5O"
+   * console.log(String(b)); // e.g. "038MQR9E14CJC12DH9DTPWFR3"
+   * console.log(String(c)); // e.g. "038MQR9E14CJC12DH9E6RJMQI"
+   * ```
+   */
+  [Symbol.iterator](): Iterator<Scru128Id, undefined> {
+    return this;
+  }
+
+  /**
+   * Returns a new SCRU128 ID object for each call, infinitely.
+   *
+   * This method wraps the result of {@link generate | generate()} in an
+   * [`IteratorResult`] object to use `this` as an infinite iterator.
+   *
+   * [`IteratorResult`]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Iteration_protocols
+   */
+  next(): IteratorResult<Scru128Id, undefined> {
+    return { value: this.generate(), done: false };
+  }
 }
 
 /** A global flag to force use of cryptographically strong RNG. */