+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Iterable and Iterator implementations to `Scru128Generator` to make it work as
   infinite iterator
 
+### Maintenance
+
+- Updated dev dependencies
+
 ## v2.1.3 - 2022-11-28
 
 ### Maintenance