+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v2.2.0 - 2022-12-25
+
+### Added
+
+- Iterable and Iterator implementations to `Scru128Generator` to make it work as
+  infinite iterator
+
+### Maintenance
+
+- Updated dev dependencies
+
 ## v2.1.3 - 2022-11-28
 
 ### Maintenance